# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZK-VOTE Relayer API",
    "version": "1.0.0",
    "description": "Backend relayer for anonymous voting on Stellar/Soroban. Generated from route definitions and Zod validation schemas — see backend/API.md for prose docs and GET /api-docs for interactive documentation."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "RelayerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relayer-Auth",
        "description": "Shared relayer auth token (also accepted as `Authorization: Bearer <token>`). There is currently one shared token per deployment, not per-user credentials."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "txHash": {
            "type": "string",
            "example": "a1b2c3...64hex"
          }
        },
        "required": [
          "success"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/bridge/vote": {
      "post": {
        "tags": [
          "Bridge"
        ],
        "summary": "Submit a cross-chain (EVM -> Soroban) vote",
        "description": "No rate limit.",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "daoId": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "proposalId": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "voteChoice": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1
                  },
                  "nullifier": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{1,64}$"
                  },
                  "voteRoot": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{1,64}$"
                  },
                  "sbtRoot": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{1,64}$"
                  },
                  "proof": {
                    "type": "object",
                    "properties": {
                      "a": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{128}$"
                      },
                      "b": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{256}$"
                      },
                      "c": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{128}$"
                      }
                    },
                    "required": [
                      "a",
                      "b",
                      "c"
                    ]
                  }
                },
                "required": [
                  "daoId",
                  "proposalId",
                  "voteChoice",
                  "nullifier",
                  "voteRoot",
                  "sbtRoot",
                  "proof"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "txHash": "a1b2c3...64hex"
                }
              }
            }
          },
          "400": {
            "description": "Error (HTTP 400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bridge/nullifier/{daoId}/{proposalId}/{nullifier}": {
      "get": {
        "tags": [
          "Bridge"
        ],
        "summary": "Check whether a nullifier has already been used",
        "description": "Rate limit: queryLimiter.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "DAO identifier",
              "example": "0"
            },
            "required": true,
            "name": "daoId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Proposal identifier",
              "example": "1"
            },
            "required": true,
            "name": "proposalId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Nullifier hash",
              "example": "0x1234..."
            },
            "required": true,
            "name": "nullifier",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "daoId": 0,
                  "proposalId": 1,
                  "nullifier": "0x1234...",
                  "used": false
                }
              }
            }
          },
          "404": {
            "description": "Error (HTTP 404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bridge/relay": {
      "post": {
        "tags": [
          "Bridge"
        ],
        "summary": "Manually trigger cross-chain event relay (admin only)",
        "description": "No rate limit.",
        "security": [
          {
            "RelayerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "success": true
                }
              }
            }
          },
          "401": {
            "description": "Error (HTTP 401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/circuits/{dao}/{type}/status": {
      "get": {
        "tags": [
          "Circuits"
        ],
        "summary": "Get the active/available ZK circuit versions for a DAO",
        "description": "Rate limit: queryLimiter.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "dao",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "comment",
                "vote"
              ]
            },
            "required": true,
            "name": "type",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "daoId": 0,
                  "circuitType": "Vote",
                  "currentCircuit": "vote_v1",
                  "availableCircuits": []
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}